fixed a problem where the wrong logger was used for measuring log entries
In the old implementation, the default play logger was instrumented to get the number of log entries written. However, since not all
log entries are written through the default logger normally, this number tended to be wrong. I now attach to the root logger which
should cover all the cases.